+++ tests/move_architecture/test_architecture.py
@@ -63,13 +63,14 @@ def test_backend_configuration_new(new_architecture):
     backend, _client = get_mocked_backend(new_architecture)
     assert backend.target.physical_qubits == [0, 1, 2, 3, 4, 5, 6]
     assert set(backend.target.operation_names) == {'r', 'id', 'cz', 'measure', 'move'}
-    assert [f'{o.name}:{o.num_qubits}' for o in backend.target.operations] == [
+    assert {f'{o.name}:{o.num_qubits}' for o in backend.target.operations} == {
         'r:1',
         'id:1',
         'cz:2',
         'measure:1',
         'move:2',
-    ]
+    }
+
     check_instruction(backend.instructions, 'r', [(1,), (2,), (3,), (4,), (5,), (6,)])
     check_instruction(backend.instructions, 'measure', [(1,), (2,), (3,), (4,), (5,), (6,)])
     check_instruction(backend.instructions, 'id', [(0,), (1,), (2,), (3,), (4,), (5,), (6,)])
@@ -85,12 +86,12 @@ def test_backend_configuration_adonis(adonis_architecture):
     backend, _client = get_mocked_backend(adonis_architecture)
     assert backend.target.physical_qubits == [0, 1, 2, 3, 4]
     assert set(backend.target.operation_names) == {'r', 'id', 'cz', 'measure'}
-    assert [f'{o.name}:{o.num_qubits}' for o in backend.target.operations] == [
+    assert {f'{o.name}:{o.num_qubits}' for o in backend.target.operations} == {
         'r:1',
         'id:1',
         'cz:2',
         'measure:1',
-    ]
+    }
     check_instruction(backend.instructions, 'r', [(0,), (1,), (2,), (3,), (4,)])
     check_instruction(backend.instructions, 'measure', [(0,), (1,), (2,), (3,), (4,)])
     check_instruction(backend.instructions, 'id', [(0,), (1,), (2,), (3,), (4,)])